Use features to prevent the util component from entraining the world in GeckoLib builds.
