+++ components/util/Cargo.toml
@@ -15,19 +15,31 @@ path = "lib.rs"
 # See https://github.com/rust-lang/rust/issues/21246
 doctest = false
 
+[features]
+
+# This feature allows us to avoid depending on various things we don't need for
+# GeckoLib builds. Conceptually, it would make more sense to have a "geckolib"
+# feature, but Cargo is generally set up for features to add dependencies, not
+# remove them. So we do it this way, and request that all non-GeckoLib builds
+# set this feature.
+non-geckolib = ["azure", "js", "layers", "html5ever", "hyper"]
+
 [dependencies.plugins]
 path = "../plugins"
 
 [dependencies.azure]
 git = "https://github.com/servo/rust-azure"
 features = ["plugins"]
+optional = true
 
 [dependencies.js]
 git = "https://github.com/servo/rust-mozjs"
+optional = true
 
 [dependencies.layers]
 git = "https://github.com/servo/rust-layers"
 features = ["plugins"]
+optional = true
 
 [dependencies.ipc-channel]
 git = "https://github.com/servo/ipc-channel"
@@ -37,7 +49,7 @@ app_units = {version = "0.1", features = ["plugins"]}
 cssparser = { version = "0.4", features = [ "serde-serialization" ] }
 log = "0.3"
 bitflags = "0.3"
-html5ever = { version = "0.2.1", features = ["unstable"] }
+html5ever = { version = "0.2.1", features = ["unstable"], optional = true }
 libc = "0.2"
 rand = "0.3"
 rustc-serialize = "0.3"
@@ -51,6 +63,6 @@ serde_macros = "0.6"
 string_cache = "0.2"
 lazy_static = "0.1"
 getopts = "0.2.11"
-hyper = "0.7"
+hyper = { version = "0.7", optional = true }
 url = {version = "0.5.2", features = ["serde_serialization"]}
 uuid = "0.1.17"
